requirements.txt
boto3>=1.28.0
botocore>=1.31.0
openai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pandas>=2.0.0
esprima>=4.0.0
semver>=3.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
beautifulsoup4>=4.12.0
requests>=2.31.0
tqdm>=4.65.0
